Fixed key events.
Made lastKeyChar() and isKeyPressed() more in line with what libraries
like Processing do. Now, lastKeyChar() is used to tell you the last
character pressed (in other words, differentiates "a" and "A") and
isKeyPressed() now accepts a string describing what key is being pressed
("a" and "A" are the same, and "SPACE" is a valid key).